Update df.py
This change will help mitigate an error raised when the number of of columns in a dataset large (search for error_bad_lines description in https://pandas.pydata.org/pandas-docs/version/0.17.0/generated/pandas.read_csv.html). I am currently facing an error in reading the opv_review table using this module in the s3-to-sheets library because of this. I believe this would fix it.